chore: move files to projects so future route can also use that convention
