+++ web/apps/dashboard/app/(app)/[workspaceSlug]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useWorkspaceNavigation } from "@/hooks/use-workspace-navigation";
-import { projectsPath } from "@/lib/navigation/routes";
+import { projectsPath } from "@/lib/navigation/routes/projects";
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
