+++ web/apps/dashboard/app/(app)/[workspaceSlug]/projects/[projectId]/_components/apps-list/index.tsx
@@ -3,7 +3,7 @@ import { ResourceCard } from "@/app/(app)/[workspaceSlug]/projects/_components/l
 import { ResourceCardSkeleton } from "@/app/(app)/[workspaceSlug]/projects/_components/list/resource-card-skeleton";
 import { useWorkspaceNavigation } from "@/hooks/use-workspace-navigation";
 import { collection } from "@/lib/collections";
-import { appDeploymentsPath, newAppPath } from "@/lib/navigation/routes";
+import { appDeploymentsPath, newAppPath } from "@/lib/navigation/routes/projects";
 import { eq, useLiveQuery } from "@tanstack/react-db";
 import { Dots, Github, Plus, Terminal } from "@unkey/icons";
 import { Button, Empty } from "@unkey/ui";
