+++ web/apps/dashboard/app/(app)/[workspaceSlug]/projects/[projectId]/apps/[appId]/(overview)/sentinel-policies/components/add-panel/forms/openapi-fields.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { projectSettingsPath } from "@/lib/navigation/routes";
+import { projectSettingsPath } from "@/lib/navigation/routes/projects";
 import Link from "next/link";
 import { useParams } from "next/navigation";
 import { Strong } from "./summary-helpers";
