+++ web/apps/dashboard/app/(app)/[workspaceSlug]/projects/[projectId]/apps/new/steps/deployment-live.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useWorkspaceNavigation } from "@/hooks/use-workspace-navigation";
-import { deploymentPath } from "@/lib/navigation/routes";
+import { deploymentPath } from "@/lib/navigation/routes/projects";
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 import { ProjectDataProvider } from "../../[appId]/(overview)/data-provider";
